+++ tests/test_mtaosCscWithSimulators.py
@@ -24,12 +24,14 @@ import unittest
 from pathlib import Path
 
 import numpy as np
+import pytest
 import yaml
 from lsst.ts import mtaos, salobj
 
 # standard command timeout (sec)
 SHORT_TIMEOUT = 5
 STD_TIMEOUT = 60
+TEST_CONFIG_DIR = Path(__file__).parents[1].joinpath("tests", "testData", "config")
 
 
 class CscTestCase(salobj.BaseCscTestCase, unittest.IsolatedAsyncioTestCase):
@@ -190,22 +192,23 @@ class CscTestCase(salobj.BaseCscTestCase, unittest.IsolatedAsyncioTestCase):
                 flush=False, timeout=SHORT_TIMEOUT
             )
 
-            # DoF after issueCorrection is rejected should be similar to DoF
-            # before applying addAberration.
+            # Aggregated DoF after issueCorrection is rejected should be
+            # similar to aggregated DoF before applying addAberration.
             self.assertTrue(
                 np.allclose(
                     dof_add_aberration_before.aggregatedDoF,
                     dof_issue_correction_after.aggregatedDoF,
                 ),
-                f"Expected {dof_add_aberration_before.aggregatedDoF} vs "
+                f"Error with aggregated DoF. Expected {dof_add_aberration_before.aggregatedDoF} vs "
                 f"Received {dof_issue_correction_after.aggregatedDoF}.",
             )
+            # Visit DoF shoud remain the same though.
             self.assertTrue(
                 np.allclose(
-                    dof_add_aberration_before.visitDoF,
+                    dof_add_aberration_after.visitDoF,
                     dof_issue_correction_after.visitDoF,
                 ),
-                f"Expected {dof_add_aberration_before.visitDoF} vs "
+                f"Error with visitDoF. Expected {dof_add_aberration_after.visitDoF} vs "
                 f"Received {dof_issue_correction_after.visitDoF}.",
             )
 
@@ -325,15 +328,137 @@ class CscTestCase(salobj.BaseCscTestCase, unittest.IsolatedAsyncioTestCase):
                 flush=False, timeout=STD_TIMEOUT
             )
 
-            assert dof_second.aggregatedDoF[0] - dof_first.aggregatedDoF[0] == 0.1
-            assert dof_second.aggregatedDoF[5] - dof_first.aggregatedDoF[5] == 0.1
-            assert dof_second.aggregatedDoF[10] - dof_first.aggregatedDoF[10] == 0.1
-            assert dof_second.aggregatedDoF[30] - dof_first.aggregatedDoF[30] == 0.1
+            assert dof_second.aggregatedDoF[0] - dof_first.aggregatedDoF[
+                0
+            ] == pytest.approx(0.1)
+            assert dof_second.aggregatedDoF[5] - dof_first.aggregatedDoF[
+                5
+            ] == pytest.approx(0.1)
+            assert dof_second.aggregatedDoF[10] - dof_first.aggregatedDoF[
+                10
+            ] == pytest.approx(0.1)
+            assert dof_second.aggregatedDoF[30] - dof_first.aggregatedDoF[
+                30
+            ] == pytest.approx(0.1)
 
             assert len(self.m2_hex_corrections) == 1
             assert len(self.cam_hex_corrections) == 1
             assert len(self.m2_corrections) == 1
             assert len(self.m1m3_corrections) == 1
+
+    async def test_stress_below_limit(self):
+        # Scenario where stress is below the limit,
+        # so no scaling or truncation should happen
+        async with self.make_csc(
+            initial_state=salobj.State.STANDBY,
+            config_dir=TEST_CONFIG_DIR,
+            simulation_mode=0,
+        ):
+            await self._simulateCSCs()
+            await self._startCsc()
+            remote = self._getRemote()
+
+            dof_aggr = np.zeros(50)
+            dof_aggr[11] = 1.0
+
+            remote.evt_degreeOfFreedom.flush()
+            await remote.cmd_offsetDOF.set_start(value=dof_aggr, timeout=STD_TIMEOUT)
+
+            updated_dof = await self.assert_next_sample(
+                remote.evt_degreeOfFreedom, flush=False, timeout=STD_TIMEOUT
+            )
+            # Assert the DOF didn't change
+            np.testing.assert_array_equal(updated_dof.aggregatedDoF, dof_aggr)
+
+            # Check final total stress is smaller or equal than the limit
+            final_total_stress = await self.assert_next_sample(
+                remote.evt_mirrorStresses
+            )
+            self.assertLessEqual(
+                final_total_stress.stressM1M3, self.csc.m1m3_stress_limit
+            )
+
+    async def test_stress_above_limit_scale(self):
+        # Scenario where stress is above the
+        # limit and sclaing approach is used
+        async with self.make_csc(
+            initial_state=salobj.State.STANDBY,
+            config_dir=TEST_CONFIG_DIR,
+            simulation_mode=0,
+        ):
+            await self.assert_next_summary_state(salobj.State.STANDBY)
+            await self._simulateCSCs()
+            remote = self._getRemote()
+
+            await remote.cmd_start.set_start(
+                configurationOverride="valid_scale.yaml", timeout=STD_TIMEOUT
+            )
+
+            await self._startCsc()
+
+            bending_stresses = self.csc.model.ofc.ofc_data.bending_mode_stresses[
+                "M1M3"
+            ]["bending_mode_stress_positive"]
+
+            dof_aggr = np.zeros(50)
+            dof_aggr[15] = 100.0
+
+            remote.evt_degreeOfFreedom.flush()
+            await remote.cmd_offsetDOF.set_start(value=dof_aggr, timeout=STD_TIMEOUT)
+
+            updated_dof = await self.assert_next_sample(
+                remote.evt_degreeOfFreedom, flush=False, timeout=STD_TIMEOUT
+            )
+
+            self.assertAlmostEqual(
+                updated_dof.aggregatedDoF[15],
+                self.csc.m1m3_stress_limit
+                / (self.csc.stress_scale_factor * bending_stresses[5]),
+                places=7,
+            )
+
+            # Check final total stress is smaller or equal than the limit
+            final_total_stress = await self.assert_next_sample(
+                remote.evt_mirrorStresses
+            )
+            self.assertLessEqual(
+                final_total_stress.stressM1M3, self.csc.m1m3_stress_limit
+            )
+
+    async def test_stress_above_limit_truncate(self):
+        # Scenario where stress is above the
+        # limit and truncation approach is used
+        async with self.make_csc(
+            initial_state=salobj.State.STANDBY,
+            config_dir=TEST_CONFIG_DIR,
+            simulation_mode=0,
+        ):
+            await self._simulateCSCs()
+            await self._startCsc()
+            remote = self._getRemote()
+
+            dof_aggr = np.zeros(50)
+            dof_aggr[10:30] = 10.0
+
+            remote.evt_degreeOfFreedom.flush()
+            await remote.cmd_offsetDOF.set_start(value=dof_aggr, timeout=STD_TIMEOUT)
+
+            updated_dof = await self.assert_next_sample(
+                remote.evt_degreeOfFreedom, flush=False, timeout=STD_TIMEOUT
+            )
+
+            # Ensure that higher-order bending modes were set to 0
+            self.assertEqual(updated_dof.aggregatedDoF[29], 0)
+            # Ensure the first element was not truncated
+            self.assertEqual(updated_dof.aggregatedDoF[10], 10.0)
+
+            # Check final total stress is smaller or equal than the limit
+            final_total_stress = await self.assert_next_sample(
+                remote.evt_mirrorStresses
+            )
+            self.assertLessEqual(
+                final_total_stress.stressM1M3, self.csc.m1m3_stress_limit
+            )
 
     async def asyncTearDown(self):
         await self._cancelCSCs()
@@ -362,24 +487,34 @@ class CscTestCase(salobj.BaseCscTestCase, unittest.IsolatedAsyncioTestCase):
 
         self.cscM2Hex.cmd_move.callback = self.hexapod_move_callbck
         self.cscCamHex.cmd_move.callback = self.hexapod_move_callbck
+        self.cscM1M3.cmd_clearActiveOpticForces.callback = (
+            self.m1m3_clear_active_optic_forces_callback
+        )
         self.cscM1M3.cmd_applyActiveOpticForces.callback = (
             self.m1m3_apply_forces_callbck
         )
         self.cscM2.cmd_applyForces.callback = self.m2_apply_forces_callbck
+        self.cscM2.cmd_resetForceOffsets.callback = self.m2_reset_force_offsets_callback
 
-    def hexapod_move_callbck(self, data):
+    async def hexapod_move_callbck(self, data):
         if data.salIndex == mtaos.utility.MTHexapodIndex.M2.value:
             self.m2_hex_corrections.append(data)
         else:
             self.cam_hex_corrections.append(data)
 
-    def m1m3_apply_forces_callbck(self, data):
+    async def m1m3_clear_active_optic_forces_callback(self, data):
+        await asyncio.sleep(1.0)
+
+    async def m1m3_apply_forces_callbck(self, data):
         self.m1m3_corrections.append(data)
 
-    def m1m3_apply_forces_fail_callbck(self, data):
+    async def m1m3_apply_forces_fail_callbck(self, data):
         raise RuntimeError("This is a test.")
 
-    def m2_apply_forces_callbck(self, data):
+    async def m2_reset_force_offsets_callback(self, data):
+        await asyncio.sleep(1.0)
+
+    async def m2_apply_forces_callbck(self, data):
         self.m2_corrections.append(data)
 
     async def _startCsc(self):